make find_mature_query in miRDeep2_core_algorithm.pl deterministic
The read with the highest frequency is used to determine the mature sequence of the candidate precursor.
If 2 or more reads have the same frequency, one of them is chosen 'by random' due to the use of hashes.
Now, the reads get additionally lexicographically ordered, making this process deterministic and replicable.