`install.pl`: Fix Perl `PDF::API2` installation
The latest version of `PDF::API2` is determined from the CPAN
`CHECKSUMS` file via a regular expression.
With the introduction of `PDF:API2::XS` this could results in installing
it instead of `PDF:API2`, which was picked up by `install.pl` as an
installation error of `PDF:API2`.
This is fixed by extending the regular expression used to match against
the (beginning) of the version number and thus avoid the clash with
`PDF:API2::XS`.

fixes #87